requirements.txt
alembic==1.16.1
aniso8601==10.0.1
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
distlib==0.3.7
et_xmlfile==2.0.0
filelock==3.12.2
flasgger==0.9.7.1
Flask==2.3.2
Flask-JWT-Extended==4.7.1
flask-marshmallow==1.3.0
Flask-Migrate==4.1.0
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.2
idna==3.4
importlib_resources==6.5.2
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
Mako==1.3.10
MarkupSafe==2.1.3
marshmallow==4.0.0
mistune==3.1.3
MouseInfo==0.1.3
numpy==2.2.5
packaging==25.0
passlib==1.7.4
pillow==11.2.1
platformdirs==3.10.0
psycopg2-binary==2.9.10
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyJWT==2.10.1
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytweening==1.2.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.31.0
rpds-py==0.25.1
six==1.17.0
soupsieve==2.7
SQLAlchemy==2.0.41
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.0.4
virtualenv==20.24.3
Werkzeug==2.3.7
wikipedia==1.4.0
